% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{ACFM}
\alias{ACFM}
\title{Average of Change in Force of Mortality with respect to lx}
\source{
Wrycza et al. (2015)
}
\usage{
ACFM(nmx, ndx, ex)
}
\description{
Average of Change in Force of Mortality with respect to lx
}
\keyword{internal}

