% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.matrix.pash}
\alias{as.matrix.pash}
\title{Convert Pace-Shape Object to Matrix}
\usage{
\method{as.matrix}{pash}(x, ...)
}
\arguments{
\item{x}{A pace-shape object.}

\item{...}{Additional arguments passed on to \code{\link{as.matrix}}.}
}
\description{
Method to convert a pace-shape object to a matrix.
}
\keyword{internal}
