% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{australia_1y}
\alias{australia_1y}
\title{Simulated Period Life-table for Australian Males 1970-1972}
\format{A data frame with 101 rows and 10 variables:
  \describe{
    \item{x}{Start of age interval in years.}
    \item{nx}{Width of age interval in years.}
    \item{nmx}{Mortality rate in age interval [x, x+nx).}
    \item{nqx}{Probability to die within age interval [x, x+nx) given survival to x.}
    \item{nax}{Subject-time spent in age interval [x, x+nx) when dying in that interval.}
    \item{lx}{Life-table survivors at age x.}
    \item{ndx}{Deaths by life-table population in age interval [x, x+nx).}
    \item{nLx}{Total person-years lived by life-table population in age interval [x, x+nx).}
    \item{Tx}{Total person-years yet to live past age x by life-table population.}
    \item{ex}{Life expectancy at age x.}
  }}
\source{
Heligman, L., & Pollard, J. H. 1980. The Age Pattern of Mortality
  Journal of the Institute of Actuaries, 107(1), 49-80.
}
\usage{
australia_1y
}
\description{
A simulated period life-table for the Australian male population years
1970-1972 in single year age intervals.
}
\details{
This life-table has been calculated from individual-level, continuous
  survival times simulated using the Heligman-Pollard mortality model with
  parameters given by Heligman and Pollard (1980), Table 2.
}
\keyword{datasets}
