% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{australia_surv_times}
\alias{australia_surv_times}
\title{Simulated Survival Times for Australian Males 1970-1972}
\format{A numeric vector with 10,000 elements.}
\source{
Heligman, L., & Pollard, J. H. 1980. The Age Pattern of Mortality
  Journal of the Institute of Actuaries, 107(1), 49-80.
}
\usage{
australia_surv_times
}
\description{
Simulated survival times for a cohort of 10,000 subjects following the
age-specific mortality rates of Australian males years 1970-1972.
}
\details{
The survival times have been simulated using the Heligman-Pollard mortality
  model with parameters given by Heligman and Pollard (1980), Table 2.
}
\keyword{datasets}
