% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{DiffAge}
\alias{DiffAge}
\title{Difference Age Vector}
\usage{
DiffAge(x, last_open = FALSE)
}
\arguments{
\item{x}{Start of age interval.}

\item{last_open}{Is the last age group open (TRUE) or closed (FALSE,
default).}
}
\value{
The widths of the age groups as a numeric vector of \code{length(x)}.
}
\description{
Get width of age groups from differenced age vector.
}
\details{
If the last age group is closed it is assumed to have the same width as the
preceeding age group. If the last age group is open, the width is set to
\code{NA}.
}
\keyword{internal}

