% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{EDaggerx}
\alias{EDaggerx}
\title{Average Life-Expectancy in Age x}
\source{
Vaupel et al. (2016)
}
\usage{
EDaggerx(nax, nx, ex)
}
\description{
Average Life-Expectancy in Age x
}
\keyword{internal}
