% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{eDaggerx}
\alias{eDaggerx}
\title{Average Years of Life Lost due to Death in Age x}
\usage{
eDaggerx(nax, nx, ex)
}
\description{
Average Years of Life Lost due to Death in Age x
}
\keyword{internal}

