% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{findvalue}
\alias{findvalue}
\title{Find and compute values that depend on ex
eg: m(x) when x = e(x)}
\usage{
findvalue(measure, x, nx, ex)
}
\description{
Find and compute values that depend on ex
eg: m(x) when x = e(x)
}
\keyword{internal}

