% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{HMD_SWE_2014}
\alias{HMD_SWE_2014}
\title{Complete life table for Swedish Females 2014}
\format{A data frame with 111 rows and 11 variables:
  \describe{
    \item{country}{Country code: SWE}
    \item{Year}{Year}
    \item{Age}{Age}
    \item{mx}{Age specific death-rates}
    \item{qx}{Age specific 1-year death probability}
    \item{ax}{ax}
    \item{lx}{Survivors at age x}
    \item{dx}{Deaths in the age interval}
    \item{Lx}{Lx}
    \item{Tx}{Tx}
    \item{ex}{Life expectancy at age x}
  }}
\source{
Human Mortality Database, \url{http://www.mortality.org}
}
\usage{
data(HMD_SWE_2014)
}
\description{
Complete life table for Swedish Females 2014
}
\keyword{datasets}

