% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pash_construction.R
\name{Inputlx}
\alias{Inputlx}
\title{Convert A Life Table Survivorship Function To A Pace-Shape Object}
\usage{
Inputlx(x, lx, nax = "midpoint", nx = "auto", last_open = FALSE,
  time_unit = "years")
}
\arguments{
\item{x}{Start of the age interval.}

\item{lx}{Probability to survive up until age x.}

\item{nax}{Subject-time alive in [x, x+n) for those who die in same interval (see details for options).}

\item{nx}{Width of age interval [x, x+n) (see details for options).}

\item{last_open}{Is the last age group open (TRUE) or closed (FALSE, default).}

\item{time_unit}{The unit of the ages (by default "years").}
}
\value{
A pace-shape object.
}
\description{
Given an age vector and corresponding survival probabilities a complete
lifetable is calculated and a pace-shape object constructed.
}
\details{
For nx you may provide a scalar, a vector numeric, or let the function
determine the width for you (default).
A scalar will be recycled for each age group. A vector must be as long as the
age vector and allows you to specify the width of each age group seperately.
If the last age group is supposed to be open make sure that the last value
of your nx vector is NA. By default the width of the age groups is guessed
from the age vector. The width of the last age group is assumed to be equal
to the width of the preceeding age group.

Likewise to nx, nax may be provided as a scalar, a vector, or automatically
derived from the width of the age groups via the midpoint method (default).
The midpoint method assumes a linear decline of the lx function over the width
of an age group, implying that those who die in that age group die on average
halfway into it. This is a common assumption in life table creation. Should
you specify an open last age group and nax estimation via the midpoint method
(not recomended) the nax of the last open age group will be set to the nax of
the preceeding age group.
}
\examples{
Inputlx(x = prestons_lx$x, lx = prestons_lx$lx)

}

