% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pash_construction.R
\name{Inputmx}
\alias{Inputmx}
\title{Convert a Life Table Mortality Rate Function to a Pace-Shape Object}
\source{
Preston, Samuel H., Patric Heuveline, and Michel Guillot. 2001.
  Demography. Oxford, UK: Blackwell.
}
\usage{
Inputmx(x, nmx, nax = "midpoint", nx = "auto", last_open = FALSE,
  time_unit = "years")
}
\arguments{
\item{x}{Start of the age interval.}

\item{nmx}{Risk of death in age interval [x, x+n)}

\item{nax}{Subject-time alive in [x, x+n) for those who die in same interval
(see details for options).}

\item{nx}{Width of age interval [x, x+n) (see details for options).}

\item{last_open}{Is the last age group open (TRUE) or closed (FALSE,
default).}

\item{time_unit}{The unit of the ages (by default "years").}
}
\value{
A pace-shape object.
}
\description{
Given an age vector and corresponding mortality rates a complete
lifetable is calculated and a pace-shape object constructed.
}
\details{
For nx you may provide a scalar, a vector numeric, or let the
  function determine the width for you (default). A scalar will be recycled
  for each age group. A vector must be as long as the age vector and allows
  you to specify the width of each age group seperately. If the last age
  group is supposed to be open make sure that the last value of your nx
  vector is NA. By default the width of the age groups is guessed from the
  age vector. The width of the last age group is assumed to be equal to the
  width of the preceeding age group.

  Likewise to nx, nax may be provided as a scalar, a vector, or automatically
  calculated via the midpoint method (default) or the constant nmx assumption
  (option "\code{constant_nmx}").

  The midpoint method assumes a linear decline of the lx function over the
  width of an age group, implying that those who die in that age group die on
  average halfway into it:

  nax = n/2 (see Preston, 2001, p. 46)

  Assuming the mortality rate during age interval [x, x+n] to be constant
  implies an exponentially declining lx function within [x, x+n] and will
  produce ax values smaller than those calculated via the midpoint method.
  Preston (2001), p. 46 provides an expression for nax given the assumption
  of constant mortality.

  nax = nx + 1/nmx - nx/(1 - exp(-nx*nmx))

  Should you specify an open last age group and nax estimation via the
  midpoint method (not recomended) the nax of the last open
  age group will be set to the nax of the preceeding age group.
}
\examples{
swe <- subset(sweden5x5, sex == "female" & period == "1940-1944")[c("x", "nmx")]
Inputmx(x = swe$x, nmx = swe$nmx, last_open = TRUE, nax = "constant_nmx")

}

