% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.pash}
\alias{is.pash}
\title{Test if Object is Pace-Shape}
\usage{
is.pash(x)
}
\arguments{
\item{x}{A pace-shape object.}
}
\description{
Method to test if an object is of class \code{pash}.
}
\keyword{internal}
