% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_measures.R
\name{LDR}
\alias{LDR}
\title{Life-table Death Rate}
\source{
Wrycza, Tomasz, and Annette Baudisch. 2014.
"The Pace of Aging: Intrinsic Time Scales in Demography."
Demographic Research 30 (1): 1571-90. doi:10.4054/DemRes.2014.30.57.
}
\usage{
LDR(ex)
}
\description{
Life-table Death Rate
}
\details{
The Life-table death rate is the inverse of the total
  life-expectancy. It can also be understood as the weighted harmonic mean of
  the force of mortality, i.e. mu-bar.
}
\keyword{internal}
