% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableCV}
\alias{LifetableCV}
\title{Life Table Coefficient of Variation}
\usage{
LifetableCV(x, ndx, nax, ex, harmonized)
}
\description{
Life Table Coefficient of Variation
}
\keyword{internal}
