% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableGini}
\alias{LifetableGini}
\title{Life Table Gini-Coefficient}
\source{
Schoeley (2017)
}
\usage{
LifetableGini(x, nax, ndx, ex)
}
\description{
Discrete formulation of the Gini-Coeffcient
}
\keyword{internal}
