% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableGini2}
\alias{LifetableGini2}
\title{Life Table Gini-Coefficient}
\usage{
LifetableGini2(nax, nx, lx, ex)
}
\description{
Another discrete formulation of the Gini-Coeffcient
}
\seealso{
LifetableGini
}
\keyword{internal}

