% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableGini3}
\alias{LifetableGini3}
\title{Life Table Gini-Coefficient}
\source{
Danko (2016)
}
\usage{
LifetableGini3(nax, nx, lx, ex, ndx)
}
\description{
Discrete formulation of the Gini-Coeffcient using
Euler-Maclaurins approximation
}
\seealso{
LifetableGini
}
\keyword{internal}

