% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableGini4}
\alias{LifetableGini4}
\title{Life Table Gini-Coefficient Variant}
\source{
Wrycza et al. (2015) - Pascariu (2016)
}
\usage{
LifetableGini4(nax, nx, lx, ex)
}
\description{
Life Table Gini-Coefficient Variant
}
\seealso{
LifetableGini
}
\keyword{internal}
