% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableVar}
\alias{LifetableVar}
\title{Life Table Variance}
\source{
Schoeley (2017)
}
\usage{
LifetableVar(x, ndx, nax, ex)
}
\description{
Discrete formulation of variance of life-table distribution of death
}
\keyword{internal}
