% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_extrapolation.R
\name{LinearExtrapolation}
\alias{LinearExtrapolation}
\title{Linear Extrapolation}
\usage{
LinearExtrapolation(x, y, xextra, loga)
}
\arguments{
\item{x}{A vector of x values.}

\item{y}{A vector of y values.}

\item{xextra}{The x for which y should be extrapolated.}

\item{loga}{Should a log-linear model be fitted?}
}
\description{
(Log)-linearly extrapolate a single (y,x) tuple.
}
\examples{
\dontrun{
LinearExtrapolation(x = 1:2, y = c(1, 2), xextra = 3, loga = FALSE)
LinearExtrapolation(x = 1:2, y = c(1, 2), xextra = 3, loga = TRUE)
}

}
\keyword{internal}

