% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_extrapolation.R
\name{LinearInterpolation}
\alias{LinearInterpolation}
\title{Linear Interpolation}
\usage{
LinearInterpolation(x0, x1, y0, y1, y)
}
\description{
Get exact x for a given y based on linear interpolation between two points.
}
\keyword{internal}
