% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pash_construction.R
\name{lxToPash}
\alias{lxToPash}
\title{Convert A Life Table Survivorship Function To A Pace-Shape Object}
\usage{
lxToPash(x, lx, nax = "midpoint", nx = "auto", last_open = FALSE,
  time_unit = "years")
}
\value{
A pace-shape object (see \code{pash}).
}
\description{
Given an age vector and corresponding survival probabilities a complete
lifetable is calculated and a pace-shape objectt constructed.
}
\examples{
lxToPash(x = prestons_lx$x, lx = prestons_lx$lx)

}

