% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{lxToPashInputValidation}
\alias{lxToPashInputValidation}
\title{Validate Input of lxToPash Function}
\usage{
lxToPashInputValidation(x, lx, nax, nx, k, last_open)
}
\description{
Validate Input of lxToPash Function
}

