% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{naxCFMfromnqx}
\alias{naxCFMfromnqx}
\title{Calculate nax From nqx Using Constant Force of Mortality Assumption}
\usage{
naxCFMfromnqx(x, nx, nqx, npx, k, last_open)
}
\arguments{
\item{nx}{Width of age interval [x, x+n).}

\item{nqx}{Probability to die within age interval [x, x+nx) given survival to
x.}

\item{npx}{Probability to survive age interval [x, x+nx) given survival to x.}

\item{k}{Number of age groups.}

\item{last_open}{Is the last age group open (TRUE) or closed (FALSE)?}
}
\value{
The nax as a numeric vector of \code{length(nx)}.
}
\description{
Get life-table nax values from nqx using the constant nmx assumption.
}
\details{
For all age groups but the last we calculate:
nax = -nx/nqx - nx/log(npx) + nx

If the last age group is closed we linearly extrapolate the last nax from the
preceding two nax.

If the last age group is open we set the last nax to NA.
}
\keyword{internal}
