% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{naxConstantnmx2}
\alias{naxConstantnmx2}
\title{nax Constant nmx (nmx)}
\usage{
naxConstantnmx2(nx, nmx, k, last_open)
}
\arguments{
\item{nx}{Width of age interval [x, x+n).}

\item{nmx}{Mortality rate in age interval [x, x+nx).}

\item{k}{Number of age groups.}

\item{last_open}{Is the last age group open (TRUE) or closed (FALSE)?}
}
\value{
The nax as a numeric vector of \code{length(nx)}.
}
\description{
Get life-table nax values from nmx using the constant nmx assumption.
}
\details{
For all age groups we calculate:
nax = nx + 1/nmx - nx/(1 - exp(-nx*nmx))

If the last age group is open we calculate the last nax as the remaining life
expectancy at x. Given the constant hazard assumption that is:
nax[k] = 1/nmx[k]
}
\keyword{internal}

