% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{naxMidpoint}
\alias{naxMidpoint}
\title{nax Midpoint}
\usage{
naxMidpoint(nx, k, last_open)
}
\arguments{
\item{nx}{Width of age interval [x, x+n).}

\item{k}{Number of age groups.}

\item{last_open}{Is the last age group open (TRUE) or closed (FALSE)?}
}
\value{
The nax as a numeric vector of \code{length(nx)}.
}
\description{
Get life-table nax values from nx using the midpoint assumption.
}
\details{
If the last age group is open then set its nax value to that of the
preceeding age group.
}
\keyword{internal}

