% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{naxUDD}
\alias{naxUDD}
\title{Calculate nax From nx Using Uniform Distribution of Deaths Assumption}
\usage{
naxUDD(nx, k, last_open)
}
\arguments{
\item{nx}{Width of age interval [x, x+n).}

\item{k}{Number of age groups.}

\item{last_open}{Is the last age group open (TRUE) or closed (FALSE)?}
}
\value{
The nax as a numeric vector of \code{length(nx)}.
}
\description{
Get life-table nax values from nx using the "uniform distribution of deaths"
(midpoint) assumption.
}
\details{
nax is NA for last open age group.
}
\keyword{internal}

