% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{nmxTonqx}
\alias{nmxTonqx}
\title{Convert Death Rates to Death Probabilities}
\usage{
nmxTonqx(nmx, nax, nx, mode)
}
\arguments{
\item{nmx}{Mortality rate in age interval [x, x+nx).}

\item{nax}{Subject-time alive in [x, x+n) for those who die in same interval.}

\item{nx}{Width of age interval [x, x+n).}

\item{mode}{Conversion formula. "chiang", "cfm", or "udd".}
}
\description{
Convert Death Rates to Death Probabilities
}
\details{
Chiangs formula for the nmx->nqx conversion is

nqx = nx*nmx / (1+(nx-nax)*nmx).

It may produce probabilities > 1 in cases where age groups are wide, hazards
are high and the nax is approximated using the uniform distribution of deaths
or the constant force of mortality assumption. Therefore in these cases we
use other conversion formulas.

nqx = 1 - exp(-nmx*nx),

assuming a constant force of mortality within the age interval [x, x+n), and

nqx = 1 - (1-nmx / (1+0.5*nmx))^nx,

assuming a uniform distribution of deaths within the age interval [x, x+n).
}
\keyword{internal}

