% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pash_construction.R
\docType{package}
\name{pash}
\alias{pash}
\alias{pash-package}
\title{pash: Pace-Shape Analysis for Life-tables}
\description{
\code{pash} lets you perform pace-shape analysis on life-tables. Calculate a
wide array of pace and shape metrics and use them to standardize and compare
your data.
}
\details{
First: Create a pace-shape object using one of the input functions:
\code{\link{Inputlx}}, \code{\link{Inputnmx}}.

Second: Calculate pace measures with \code{\link{GetPace}}, calculate shape
measures with \code{\link{GetShape}}, standardize a life table by pace and
shape with \code{\link{StandardizeLT}}, rebase your life table to a different
age with \code{\link{RebaseLT}}.
}

