% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{prestons_lx}
\alias{prestons_lx}
\title{Period Life-table Survival Function for Austrian Males 1992}
\format{A data frame with 19 rows and 2 variables:
  \describe{
    \item{x}{Start of age interval in years.}
    \item{lx}{Life-table survivors at age x.}
  }}
\source{
Preston, Samuel H., Patric Heuveline, and Michel Guillot. 2001.
        Demography. Oxford, UK: Blackwell, p. 51 and United Nations 1994.
}
\usage{
prestons_lx
}
\description{
An abridged period life-table lx column for Austrian males in 1992 used in
an example on life-table construction in Preston's demography textbook.
}
\keyword{datasets}
