% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.pash}
\alias{print.pash}
\title{Print Pace-Shape Object}
\usage{
\method{print}{pash}(x, radix = 10000, big.mark = ",", ...)
}
\arguments{
\item{x}{Pace-shape object.}

\item{radix}{Initial life-table population.}

\item{big.mark}{See \code{\link{formatC}}.}

\item{...}{Additional arguments passed on to \code{\link{print.data.frame}}.}
}
\description{
Printing method for a pace-shape object
}
\keyword{internal}

