% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_transform.R
\name{RebaseLT}
\alias{RebaseLT}
\title{Rebase Life Table to Different Age}
\usage{
RebaseLT(pash, origin = 0)
}
\description{
Rebase Life Table to Different Age
}
\examples{
pash = lxToPash(x = prestons_lx$x, lx = prestons_lx$lx)
RebaseLT(pash, 50)

}

