% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.pash}
\alias{summary.pash}
\title{Summarize Pace-Shape Object}
\usage{
\method{summary}{pash}(object, ...)
}
\arguments{
\item{object}{A pace-shape object.}

\item{...}{Additional arguments.}
}
\description{
Summary method for the pace-shape object.
}
\keyword{internal}
