% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{ValidateLastnqx}
\alias{ValidateLastnqx}
\title{Validate / Modify Last nqx To Close The Life-table}
\usage{
ValidateLastnqx(x, nqx, nax, nx, last_open)
}
\description{
Validate / Modify Last nqx To Close The Life-table
}
\keyword{internal}
