% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{ValidateTaillx}
\alias{ValidateTaillx}
\title{Validate / Modify lx Tail To Close the Life-table}
\usage{
ValidateTaillx(x, lx, nax, nx, last_open)
}
\description{
Validate / Modify lx Tail To Close the Life-table
}
\keyword{internal}
