% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pash_construction.R
\name{ConstructPash}
\alias{ConstructPash}
\title{Pace-Shape Object Constructor}
\usage{
ConstructPash(x, nx, nmx, nax, nqx, npx, lx, ndx, nLx, Tx, ex, last_open,
  time_unit, type, input)
}
\description{
Pace-Shape Object Constructor
}

