% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pash_construction.R
\name{ConstructPash}
\alias{ConstructPash}
\title{Pace-Shape Object Constructor}
\usage{
ConstructPash(x, nx, nmx, nax, nqx, npx, lx, ndx, nLx, Tx, ex, last_open,
  nax_mode, time_unit, type, input)
}
\arguments{
\item{x}{Start of age interval.}

\item{nx}{Width of age interval [x, x+nx). Must align with x. No zero width
allowed.}

\item{nmx}{Mortality rate in age interval [x, x+nx) measured in deaths per
unit of subject-time.}

\item{nax}{Subject-time spent in age interval [x, x+nx) when dying in that
interval.}

\item{nqx}{Probability to die within age interval [x, x+nx) given survival to
x. Must be in range [0, 1].}

\item{npx}{Probability to survive age interval [x, x+nx) given survival to x.
A numeric vector. Must be in range [0, 1].}

\item{lx}{Life-table survivors at age x. Radix must be 1. Must be in range
[0, 1].}

\item{ndx}{Deaths by life table population in age interval [x, x+nx).}

\item{nLx}{Total subject-time lived by life-table population in age interval
[x, x+nx).}

\item{Tx}{Total subject-time yet to live past age x by life-table population.}

\item{ex}{Life expectancy at age x.}

\item{last_open}{Is the last age group open? Boolean scalar.}

\item{nax_mode}{The nax specification provided by the user. String scalar.}

\item{time_unit}{The unit of the ages. String scalar.}

\item{type}{The type of input used to create the pace-shape object. String
scalar.}

\item{input}{The raw input arguments of the Input[*] function. A list.}
}
\value{
An S3 object of class \code{pash}.
}
\description{
Construct a pace-shape object.
}
\details{
Unless otherwise noted in the argument description the input
arguments must be non-negative numeric vectors of length > 1 and length ==
length(x). NA, NaN, Inf, -Inf are not allowed.

Argument \code{type} must be one of \code{c("lx", "nmx")}. Argument
\code{nax_mode} must be one of \code{c("midpoint", "constant_nmx",
"vector", "scalar")}.
}
\section{Warning}{

Only use when writing new Input functions. For pace-shape object construction
use the existing Input functions.
}
\keyword{internal}

