% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_measures.R
\name{GetPace}
\alias{GetPace}
\title{Life Table Pace Measures}
\source{
Wrycza, Tomasz, and Annette Baudisch. 2014.
"The Pace of Aging: Intrinsic Time Scales in Demography."
Demographic Research 30 (1): 1571-90. doi:10.4054/DemRes.2014.30.57.
}
\usage{
GetPace(pash, type = "all", q = 0.5)
}
\arguments{
\item{pash}{A pace-shape object.}

\item{type}{Which pace measure should be returned (see details for choices)?}

\item{q}{(optional) Quantile specification for age where q percent of the
life table population is still alive (default 0.5).}
}
\description{
Get life table shape measures from pace-shape object.
}
\details{
The type argument accepts the following strings:
\describe{
  \item{\code{"e0"}}{Total life expectancy.}
  \item{\code{"qlx"}}{Age where q percent of the life table population is still alive.}
  \item{\code{"maxx"}}{Maximum lifespan.}
  \item{\code{"all"}}{All of the above measures}
}
}
\examples{
pash = Inputlx(x = prestons_lx$x, lx = prestons_lx$lx)
GetPace(pash)

}

