% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_measures.R
\name{GetPace}
\alias{GetPace}
\title{Life-table Pace Measures}
\source{
Wrycza, Tomasz, and Annette Baudisch. 2014.
"The Pace of Aging: Intrinsic Time Scales in Demography."
Demographic Research 30 (1): 1571-90. doi:10.4054/DemRes.2014.30.57.
}
\usage{
GetPace(pash, type = "all", q = 0.5)
}
\arguments{
\item{pash}{A pace-shape object.}

\item{type}{Which pace measure should be returned (default \code{"all"})?}

\item{q}{Quantile specification for age where q percent of the life-table
population is still alive (defaults to median).}
}
\description{
Get life-table shape measures from pace-shape object.
}
\details{
The type argument accepts the following strings:
\describe{
  \item{\code{"e0"}}{Total life expectancy.}
  \item{\code{"qlx"}}{Age where q percent of the life table population is still alive.}
  \item{\code{"ldr"}}{The life-table death rate.}
  \item{\code{"all"}}{All of the above measures}
}
}
\section{l(x) quantile interpolation}{

Exact values for the quantiles are determined by either linear- or
exponential interpolation of the l(x) function. Given a value q we first find
the x and l(x) at the boundaries of the interval containing l(x) = q. Then we
return the x where l(x) = q by either:
\describe{
  \item{linear interpolation}{In case of nax = \code{"scalar" | "vector" |
  "udd"} then x = (-x1*q+x0*q-x0*y1+x1*y0)/(y0-y1).}
  \item{exponential interpolation}{In case of nax = \code{"cfm"}
  then x = ((x0-x1)*log(y)-x0*log(y1)+x1*log(y0))/(log(y0)-log(y1))}
}
}

\examples{
pash <- Inputlx(x = prestons_lx$x, lx = prestons_lx$lx)
GetPace(pash)

# the age only 30\% of the life-table population reaches
GetPace(pash, q = 0.3)
# the estimate changes with changing nax assumptions
pash <- Inputlx(x = prestons_lx$x, lx = prestons_lx$lx, nax = "cfm")
GetPace(pash, q = 0.3)
}
