% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{GetShape}
\alias{GetShape}
\title{Lifetable Shape Measures}
\usage{
GetShape(pash, type = "all")
}
\arguments{
\item{pash}{Object of class \code{pash}.}
}
\description{
Get lifetable shape measures from pace-shape object.
}
\examples{
pash = lxToPash(x = prestons_lx$x, lx = prestons_lx$lx)
GetShape(pash)

}

