% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{GetShape}
\alias{GetShape}
\title{Life Table Shape Measures}
\source{
Wrycza, Tomasz F., Trifon I. Missov, and Annette Baudisch. 2015.
"Quantifying the Shape of Aging." PLOS ONE 10 (3): 1-18. doi:10.1371/journal.pone.0119163.

Vladimir, Shkolnikov M. and Evgheny, Andreev M. 2010.
"Spreadsheet for calculation of life-table dispersion measures". Demographic Research.
\url{http://www.demogr.mpg.de/papers/technicalreports/tr-2010-001.pdf}
}
\usage{
GetShape(pash, type = "all")
}
\arguments{
\item{pash}{A pace-shape object.}

\item{type}{Which shape measure should be returned (see details for choices)?}
}
\description{
Get life table shape measures from pace-shape object.
}
\details{
The type argument accepts the following strings:
\describe{
  \item{\code{"LTentr"}}{Life table entropy.}
  \item{\code{"LTgini"}}{Life table Gini coefficient - Pascariu implementation.}
  \item{\code{"LTgini2"}}{Life table Gini coefficient - Shkolnikov implementation.}
  \item{\code{"LTcv"}}{Life table coefficient of variation.}
  \item{\code{"all"}}{All of the above measures.}
}
}
\examples{
pash = Inputlx(x = prestons_lx$x, lx = prestons_lx$lx)
GetShape(pash)

}

