% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{Inputlx}
\alias{Inputlx}
\title{Convert a Life-table Survivorship Function into a Pace-Shape Object}
\source{
Preston, Samuel H., Patrick Heuveline, and Michel Guillot (2001).
  Demography: Measuring and modeling population processes. Oxford: Blackwell.
}
\usage{
Inputlx(x, lx, nax = "udd", nx = "auto", last_open = FALSE,
  time_unit = "years")
}
\arguments{
\item{x}{Start of the age interval.}

\item{lx}{Probability to survive up until age x.}

\item{nax}{Subject-time alive in [x, x+n) for those who die in same interval
(either numeric scalar, numeric vector or one of \code{c("udd", "cfm")}).}

\item{nx}{Width of age interval [x, x+n) (either numeric scalar, numeric
vector or \code{"auto"}).}

\item{last_open}{Is the last age group open (TRUE) or closed (FALSE, default)?}

\item{time_unit}{The unit of the ages (by default "years").}
}
\value{
A pace-shape object.
}
\description{
Given an age vector and corresponding survival probabilities a complete
life-table is calculated and a pace-shape object constructed.
}
\section{nx handling}{
 For nx you may provide either a numeric scalar, a
  numeric vector, or let the function determine the width for you
  (\code{"auto"}, default). A scalar will be recycled for each age group. A
  vector must be as long as the age vector and allows you to specify the
  width of each age group separately. By default, the width of the age
  groups are calculated from differencing the age vector.
}

\section{nax handling}{
 nax may be provided as either a numeric scalar, a
  numeric vector, or calculated via the uniform distribution of deaths
  (\code{udd}) method (default) or the constant force of mortality assumption
  (option \code{cfm}).

  The (\code{udd}) method assumes a linear decline of the l(x) function over
  the width of an age group, implying that those who die in that age group
  die on average halfway into it (also known as the "midpoint" assumption):

  nax = n/2 (see Preston et al. 2001, p. 46)

  Assuming the mortality rate during age interval [x, x+n) to be constant
  (\code{cfm} method) implies an exponentially declining l(x) function
  within [x, x+n) and will produce nax values smaller than those calculated
  via the \code{udd} method. Preston et al. (2001, p. 46) provide an expression
  for nax given the assumption of constant mortality. Restating this expression
  in terms of nqx and npx leads to:

  nax = -n/nqx - n/log(npx) + n

  If the last age group is open and the \code{udd} or \code{cfm} method is
  used, then the last nmx value is log-linearly extrapolated based on the
  preceding two nmx and the nax, and the ex for the last age group id
  calculated using the constant hazard assumption.
}

\examples{
Inputlx(x = prestons_lx$x, lx = prestons_lx$lx)
# open last age group
Inputlx(x = prestons_lx$x, lx = prestons_lx$lx, last_open = TRUE)
# different nax assumptions
Inputlx(x = prestons_lx$x, lx = prestons_lx$lx, nax = "cfm")

}
