% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableGini}
\alias{LifetableGini}
\title{Life Table Gini-Coefficient}
\usage{
LifetableGini(nax, nx, lx, ex)
}
\description{
Life Table Gini-Coefficient
}
\seealso{
LifetableGini2
}
\keyword{internal}

