% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableGini}
\alias{LifetableGini}
\title{Life Table Gini-Coefficient}
\source{
Pascariu (2016)
}
\usage{
LifetableGini(nax, nx, lx, ex)
}
\description{
Discrete formulation of the Gini-Coeffcient
}
\seealso{
\code{\link{LifetableGini}}
}
\keyword{internal}
