% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableGini2}
\alias{LifetableGini2}
\title{Life Table Gini-Coefficient (Shkolnikov)}
\usage{
LifetableGini2(nax, nx, lx, ex)
}
\description{
Life Table Gini-Coefficient (Shkolnikov)
}
\seealso{
LifetableGini
}
\keyword{internal}

