% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_measures.R
\name{LifetableVar}
\alias{LifetableVar}
\title{Life Table Variance}
\source{
Pascariu (2016)
}
\usage{
LifetableVar(nax, nx, ex, ndx)
}
\description{
Discrete formulation of variance
}
\keyword{internal}

