% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_transform.R
\name{StandardizeLT}
\alias{StandardizeLT}
\title{Standardize Life-table By Pace}
\source{
Wrycza, Tomasz F., Trifon I. Missov, and Annette Baudisch. 2015.
  "Quantifying the Shape of Aging." PLOS ONE 10 (3): 1-18.
  doi:10.1371/journal.pone.0119163.
}
\usage{
StandardizeLT(pash, pace = "e0", q = 0.5)
}
\arguments{
\item{pash}{A pace-shape object.}

\item{pace}{Measure of pace used for standardization (default e0).}

\item{q}{Quantile specification for age where q percent of the life-table
population is still alive (defaults to median).}
}
\value{
A pace standardized life table in form of a data frame. Note
that this is not a pace-shape object anymore.
}
\description{
Standardize Life-table By Pace
}
\examples{
# generate a pace-shape object
pash = Inputlx(x = prestons_lx$x, lx = prestons_lx$lx)
# standardize the life table by pace and shape
StandardizeLT(pash)
StandardizeLT(pash, pace = "qlx", q = 0.9)

}
