% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_transform.R
\name{StandardizeLT}
\alias{StandardizeLT}
\title{Standardize Life Table By Pace and Shape}
\usage{
StandardizeLT(pash)
}
\description{
Standardize Life Table By Pace and Shape
}
\examples{
pash = lxToPash(x = prestons_lx$x, lx = prestons_lx$lx)
StandardizeLT(pash)

}

